Refactor: Extract HTTP client and CSV parser utilities
